refactor: setup use-carousel hooks
